src-python/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pyinstaller>=6.10.0
pytest>=7.4.3
httpx>=0.25.2
requests>=2.31.0
# pywebview 桌面壳
pywebview>=5.0.0
# 剪贴板支持
pyperclip>=1.8.0